requirements.txt
asgiref==3.10.0
crispy-bootstrap5==2025.6
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.8
django-crispy-forms==2.5
django-import-export==4.3.14
pillow==12.0.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
tablib==3.9.0
tzdata==2025.2
Unidecode==1.4.0
